Allow users to transfer matches between facilities
In some cases, list items have been incorrectly matched to a facility.
Currently, users will 'split' those list items away from the incorrect
facility. As a second step, they sometimes merge the match to a
different facility.

However, when the list items are not able to be geocoded, it's
impossible to use the 'split' functionality to create a new facility
from the match.

This adds the functionality to move or transfer a list item directly
from one existing facility to another existing facility, avoiding the
need to create an new, intermediate facility.