Add Geocoding Dashboard
Several facility list items have been either geolocated to an incorrect
location or have failed to geocode. Some of these addresses have been
successfully searched on Google maps directly, suggesting that the
issues are either sporadic or are related to the combination of
country code and address used to geolocate the facility. In order to
more easily explore and debug these issues, we have added a direct
geocoding endpoint and associated administrative dashboard page.