Fix delete facility endpoint
When deleting a facility with multiple confirmed or automatic matches,
we attempt to split off and create a new facility from the top match.
In the case that the match was missing a location, this caused a null
location error, preventing the facility from being deleted.

Matches without locations are now filtered out when searching for valid
matches to use to create a new facility during the deletion process.