Redirect to alias for deleted facilities
When a facility is not found, determines if any aliases exist for the
given OAR ID, then redirects to the first alias found, only returning
a 404 if no aliases exist.

Note that the redirect causes the browser to automatically refetch the
data from the provided redirect endpoint, and Axios doesn't have
access to interrupt the response and redirect to a new webpage. However,
once the data is fetched, in the facility details component we check
if the returned data has a OAR ID patching the current URL parameters.
If not, we redirect the browser to the alias.