Add linking IDs for relocated facilities
Sometimes facilities close and relocate to a new address. In these
instances, the original facility will be marked as closed, but we
should be able to link the closed facility to the new facility ID.

Unlike deleted facilities, users should still be able to see the closed
facility's details, so they are not rerouted automatically. Instead, on
a closed facility profile, a link is added to allow users to easily
navigate to the new facility details page.

In addition, administrators require a method to link the closed
facilities to the new ids. Therefore a new page has been added to the
admin dashboard allowing a new OAR ID to be linked to a facility.