Fix invalid country/address combinations
Three facilities contributed to the OAR have addresses in China but,
erroneously, have India listed as the country in the data upload.
These facilities have plotted on the map in China, but appear in India
country searches because of their assigned country code.

Reviewing the raw data for the associated facility
list items, it appears that the submitted country code was for India,
although the addresses appear to be in China. This seems like an error
in the submitted data.

To prevent similar invalid facilities from being inserted into the
database in the future, we now raise a geocoding error when the geocoded
location's country code doesn't match the provided country code.