Add warning about inexact facility coordinates
Sometimes facilities are geolocated incorrectly or are unable to be
geolocated. In that situation, admins will manually update the location;
however, we need to alert the users that the facility coordinates are
known to be inaccurate. In the Django admin, users are now also able to
note that the facility is known to be inaccurate. When coordinates have
been labeled as inexact, we add text to the facility details noting that
the location is known to be inexact.