Allow blank new_oar_ids
A new OAR ID, linking a facility to a new location / ID, should not be
required, because most facilities have not been moved. However, the
Django admin was requiring users to enter a new ID in order to save
changes to facilities. A migration has been added to update the model
to allow blank new_OAR_IDs.