Add endpoints for info website
In the info website relaunch, there will be 3 stats listed on the home
page about the OAR: facilities count, active contributors count, and
count of countries with active facilities. In order for those numbers
to automatically update as the OAR updates, we have added endpoints to
provide the active contributors count and the active countries count.