Update header and footer
The info website has been updated with a new header and footer. In
order to keep users' experiences as similar as possible between
the info site and the map site, updates the header and footer to match.

A code sample was provided for the info website's header and footer,
and wireframes were provided for the full-width view of the map site's
header and of the info site's mobile header. As much consistency as
possible with these wireframes has been used. Differences include
adding authentication, translation, and legal links to the mobile
navigation, and adding the API documentation link to the full-width
footer.