Enable cross-site access for count endpoints
The info website needs to be able to access the `count` endpoints in
the backend, but a combination of CORS and general permissions has been
blocking its access. Enables CORS for GET requests to /api/info
endpoints from certain origins, and allows those endpoints to be
accessed without the ordinary protections.