Add ExtendedFields model
Adds the ExtendedFields model, the first step towards extended profiles.

The `field_name`, describing the field name for the extended field,
is pulled from a list of choices. Currently, available field_names are
country, address, name, number_of_workers and native_language_name.

The `value` is stored in a jsonb object, with different structure for
different field types. Numeric fields will be stored as {“min”: v1,
“max”: v2}. If there is a single numeric value, we will set both min
and max to it. If there is no provided value, we won't make an
ExtendedField row.

Each ExtendedField will be created from either a facility claim or a
facility list item. It will be must linked to one or the other, but
not both.