Fix 'Processing' links
The links to https://info.openapparel.org/processing on the 'My Lists'
and 'Contribute' pages weren't updated to use external links following
moving the processing page to the info site, causing React-Router to
'stack' the links (adding the full processing url on top of the current
page's url). Switching to anchor tags resolves this issue.