Correct info site routes
Adds the info site routes as constants in the constants file,
instead of using their string value directly. This allows us to
update the links in one place in order to keep them consistent and
avoid outdated links around the site.

Updates the info links throughout the site to use the constant values.

Updates the info site routes to use the most current version of the
URLs.