Make ExtendedField.Facility nullable
The ExtendedFields from FacilityListItems will be created in two steps:
firstly, they will be initialized with a null facility field when
parsing the list item, at which time the list item doesn't have an
assigned facility; secondly, they will be assigned to a facility in the
matching phase, when the list item is linked to a facility.

To allow the first step, the facility field is now nullable.