Update extended fields when matches are adjusted
When merging two facilities, splitting a facility list item into its
own facility, or transferring a list item from one facility to another,
the list items in question are adjusted to point to a new facility.

ExtendedFields which are created from a list item contain both a
reference to the list item and a reference to the facility. In order to
keep the fields and list item in sync, we now update the ExtendedFields
when the list item's facility is changed.