Fix broken geocoding test
When geocoding a facility, if the country code returned by Google Maps
is different from the country code provided by the user, we throw a
ValueError.

We’re testing this code with an address that previously was getting
mapped by Google Maps to China, and testing that we threw an error
because the user had entered it as India. Google is now returning that
address as being in India, which means that a value error was
(correctly, but breaking the test) not raised.

The test now mocks the API response so that we are testing only our
own code, and not third-party responses.