Use the contributor's facility name in embed
In the EmbeddedMap, the facility name (not the address) of the facility
that was contributed by the embedded map's contributor is
promoted to the primary position. This is only be in effect in the
embedded map view, and not in the full site. If the contributor has
multiple active list items for the facility, we prefer first any that
match the assigned facility name, or secondly the most submission.