Fix contributor id bug
When getting the contributor id for the facilities list from the
request params `contributors`, we were incorrectly grabbing it as a
string instead of a list. Then, when accessing the first item, the
first digit of the contributor's id was selected instead of the first
contributor id. This resulted in the wrong contributor being selected
when the contributor had a multi-digit id. Now we parse the param as
a list.