Handle create=false when exact matching
When a facility list item is submitted with create=false and there is
no match found, a FacilityListItem is created with facility_id=None.
When another facility list item is submitted which exact matches that
FacilityListItem, an error was being thrown. Instead, we now filter
out FacilityListItems where the facility_id=None, allowing us to match
to the next best option (if one exists) or pass the FacilityListItem
on to the dedupe process.