Fix ExtendedField admin
Loading an individual ExtendedField in the Django admin on staging
causes a server timeout. It seems likely that this is due to the
automatic setup of the admin, which creates dropdown lists for
contributor, facility, facility claim, and facility list item, some of
which include thousands of items. Because we are unlikely to want to
update these fields in the admin (rather, we would create a new
extended field), I have updated them to be readonly, which prevents the
fetching of so much data and should resolve the timeout errors.