Prevent `clean_` field error when parsing
When there is an error parsing a facility, we attempt to create a
list item with an 'ERROR PARSING' status. If this list item has an
empty address or name field, it threw an error due to the not-null
rule for clean_name and clean_address. If a `clean_` field would get
set to `None`, instead set them to an empty string.