Add ContributorManager with filter_by_name method
To support fuzzy matching contributors by a submitted parent company extended
field we add a custom manager to the Contributor model.

We set the TRIGRAM_SIMILARY_THRESHOLD by trial and error testing with
development data.

In the real world multiple people register accounts with the same contributor
name to we prioritize matches to a verified account, then to one under which
data has been submitted.