Index custom text
Contributors want to mark specific contributor fields / embed fields to
be included in the text which is searched by the facilities query.
This adds the function to update the `custom_text` field in the
FacilityIndex, which will allow those custom fields to be included in
the facilities search.

Whenever the FacilityIndex is updated, and in post_save for
FacilityListItems, and when updating a user's embed config, the
index_custom_text function is passed an array of facility ids to
update.

We get a list of contributors who have searchable embed configs,
and for each of the facilities in the facility_ids list, find the most
recent FacilityListItem submitted for that facility by each contributor,
if one exists. Then we grab any fields from the `raw_data` of that
facility which are included in the submitting contributor's searchable
fields, and concatenate them into a single text value. Finally, we
update each facility's `custom_text` to the calculated combined text
value.