Handle facility/processing type when processing new facilities
The field values are used as look-up values in the facility/processing
type matching function. Single and multiple types are handled from both
the API endpoint and when processing a CSV. If any of the submitted
types are unabled to be matched, a parsing error is returned for that
facility.

Refs #1583