Handle country code geocoding errors
Submitted facilities are frequently throwing the country code error,
which indicates that the geocoding process has returned a result that
doesn't contain the country which the contributor has submitted for the
facility.

There are two primary instances where this is occurring:

1) The returned geocoding result is imprecise and contains no country
code at all. In this case, we will now accept this as a valid result.

2) There are multiple results returned, and although the top result
has an invalid code, another result contains the correct country code.
In this case, we will now select the result with the correct country
code. (If there are both a result with a correct country code and an
inexact result, we prefer the correct country code.)

We will still return a geocoding error if all results include
non-matching country codes.