Rename embed default fields to extended
Prior to the general availability of extended fields embedded maps had special
handling for transparency pledge fields. As part of removing this special
handling we rename "default" to "extended" and update the field keys to aligh
with the extended field keys.

Note that although we have implemented support for native language name it is
not part of the publicly acknowledged lists of extended fields that will be
available when the extended profile feature is launched. As a result, we do not
included it in `NonstandardField.EXTENDED_FIELDS`.

Adding `processing_type` to the list of nonstandard fields required us to update
a couple of unit test assertions that count the number of fields.