Filter extended fields by contributor
In order to show extended fields on an embedded map we need the ability to
filter them so that only values added by the embedded map contributor appear.

We also remove handling for extended fields from the `get_contributor_fields`
serializer method because the values are now retrieved as by the
`get_extended_fields` serializer method.