Update embed facility detail rendering to support extended fields
Embedded map owners have the ability to change the order of and labels for the
fields they have submitted. We add support for this to extended fields and
support for mixing the order of extended fields and other contributor fields by
introducing a new `renderEmbedFields` method that references the embed config
and calls the appropriate function to render either an extended field or an
arbitrary contributor field.

When showing an embedded map we only show the most recent extended field values
from the embed contributor. As such, we don't need to display the secondary line
of information under the field value since it will always display the same
contributor name.