Include other_ fields in ExtendedFields
Includes the other_names and other_addresses
information with the extended_fields when serializing, to allow proper
server-side sorting and inclusion of contributor data.

Prioritizes the name from an active/verified facility claim over the
default facility object's name when serializing the Facility.name,
(unless it is in embed mode and prefer_contributor_name is active, in
which case prefers the contributor name).

Adjusts the created_from to hide other contributor information in embed
mode.