Update state when leaving facility details
When navigating away from a facility, we need to clear the facility
currently in state, or the facility redirect process identifies the
stale facility ID as an alternate id and redirects to the stale
facility. This can result in an infinite loop.

The same issue was causing a redirect loop when navigating to alternate
facilities by clicking on the map while in the facility details.

We were also not correctly clearing the facility and refetching the
search results when navigating away from the facility and back to the
facilities list.