Ignore partial OAR ID matches in text search
When searching for a number, such as '12', in the OAR text search, all
facilities with the number '12' in their OAR ID will be returned. In
general, users who want to match to an OAR ID will search for the
entire ID, and not just a portion of the ID. This results in many more
facilities being returned for numeric searched than users are expecting.

We now only consider the query text to be a match to the OAR ID if it's
an exact match.