Prevent navigation loop for single-result queries
When the search is run for the facilities list, if there is only one
match, the search bar automatically navigates to the facility. However,
when we navigate back to the list from the facility details, this
resulted in being navigated directly back to the facility details.

When rerunning the search in the back navigation for facility details,
replace the 'push' command with a noop to prevent navigation to a
specific facility.