Temporarily Fix Google Basemap
The API version to which we have manually set our Google basemap
has been deprecated for some time. When the requested version is
unavailable, the API defaults to returning the most current version
(currently, '3.48.1a', which can be seen by entering
'google.maps.version' in the app console).

The Google basemap is currently not loading in staging or production.
Although release notes for 4.8 have not been released, it was scheduled
to release around this time, so it seems likely that the release of
4.8 has some breaking change effecting our basemap.

Manually setting the version to 4.7 (a previous but still-available
version) results in the basemap loading as expected.

Version 4.7 should be available for at least the next quarter (per
https://developers.google.com/maps/documentation/javascript/versions,
it should be available for 3-6 months), which will give us time to
update our code to fix the basemap on current versions.