Fix handling of numeric cells in Excel workbooks
With the introduction of number_of_workers we now have a column that can have a
numeric value. This presented 2 challenges to the existing Python code.

- Excel parsing library returns these as numeric types, but the code was
expecting a string.
- A value that appears as 130 in the workbook can appear as a float 130.0 after
parsing.

This commit attempts to fix both of these issues by first manually casing cell
values to strings when parsing the workbook, then using an `int(parse(x))`
nested cast in the `extract_int_range_value` function to ensure that the decimal
portion is removed.