Modify custom_text handling
Contributors want search queries in their embedded map to only return
matches for custom text if it is an exact match to custom text which
they submitted.

We have updated the custom_text field to use an array of strings in the
format '{contributor_id}|{custom_text}' (For example, '1|abcd'). We
format the free text query in the search in the same way, and find exact
matches.

In our previous iteration of this feature, given Contributor 1 submitted
a custom text value of "ABCD" for Facility A and Contributor 2
submitted a custom text value of "ABCD" for Facility B, searching for
"D" in Contributor 1's map could return both Facility A and Facility B.

In the current version, searching "D" would return no matches; and
searching "ABCD" in Contributor 1's map would return only Facility A.