Add support for facility_type_processing_type meta extended field
In an attempt to avoid confusion the data submission process we will be putting
a single `facility_type_processing_type` column in our upload templates rather
than separate `facility_type` and `processing_type` columns. Though we are
unifying the fields in the template we still want to create separate extended
fields for each field type. Supporting this is a simple change because our
processing was designed to handle either value type in either field.