Combine facility type and production type claim fields
In an effort to simplify the contribution process the OAR team has decided, in
CSV/Excel file templates and the facility claim form, to unify the facility type
and processing type fields. To accomplish this on the clame form we remove the
facility type field from the UI and on the backend create extended fields for
both facility type and processing type based on the values selected for
processing type.