Include extended fields in CSV/Excel download
Adds extended fields values to the CSV/Excel downloads.

The fields are included as number_of_workers, parent_company,
facility_type, facility_type_raw, processing_type, processing_type_raw,
and product_type. Because the matched values might not be significantly
different from the submitted values for facility_type and
processing_type (due to aliasing), the raw values are included in the
download by client request.

Multiple values in a cell are combined by separating them with a
vertical bar, and each value include the contributor name and date:
Value (Contributor Name on Jun 1 2021).

In embed mode, the contributor name and the date are not included,
and only the values are shown. Values submitted by contributors
other than the owner of the embedded map are not included in the
download.