Add embed_level to development contributors
We frequently need to test embedded map settings and embedded map mode.
If `resetdb` has been run, this requires logging in as an admin and
giving embed permissions to one of the development contributors.

Adds a default embed_level of 3 to c2@example.com, 2 to c3@example.com,
and 1 to c4@example.com to avoid needing to manually set these
permissions.