Make development contributor and user ids differ
In development, the default contributor and user IDs match. This has
caused us to miss a number of bugs related to mixing up contributors
and users when working with data (because these two entities are used
somewhat interchangeably). The IDs of contributors and their associated
users now will no longer match after running `resetdb`, which will
hopefully prevent this category of bug.