Fix claim showing incorrect parent facility
The claim facility form was correctly submitting and saving the
contributor name and contributor id; however, when serializing the data
for fetching it, we were incorrectly grabbing the user id. This caused
the form to display the incorrect contributor in the dropdown.

Claims now correctly serialize the parent_facility with the contributor
id, not the admin user id.