Add an API to return suggested product type choices
We allow contributors to submit any value for product type but expect that by
suggesting standard values and values that have already been submitted we will
have more standardization in the collected data.

We use a helper function so we can share the logic with a public API and with
the claim form serializer which was already pulling a list of values. Now that
we have migrated claims to use extended fields we can replace the old logic used
to look for product types used on other claims.