Add parent_company "creatable" multi-select list
Use react-select “creatable” to specify one ore more contributors.
The data source for the select list is the same as the "Filter by
Contributor" field.
If one of these is selected, set the query string argument to the
numeric ID; If the item is not in the list, set the query string
argument to the text name.
Update list with label to make sure parent company name instead of id
shows up when accessing through url