Add facility_type and processing_type multi-select list
Make request to fetch the facility/processing type choice values, and
parse the response to populate the facility type and processing type
multi-select list.
The two multi-select lists are connected - If one or more facility_types
is selected, only allows selecting processing types that belong to one
of those types, and vice versa.