Move exact_match sorting to server
When a pending facility list item has multiple exact matches in the
database, we choose the match to automatically pair with via sorting
the matches and choosing the top result. We prefer active items, then
items which were submitted by the current contributor, then by the
most recent submission.

Previously, we sorted exact_matches using Django annotations and
order_by. This included a CPU-intensive nested subquery to collect a
list of active list item ids.

Moving the sorting to a server-side
sort allows us to perform the subquery only once, outside of the main
query, reducing the time taken by the query.