Use CreatableInputOnly for product type search
The OAR team decided to remove the product types dropdown list since we accept
arbitrary values from contributors and they can be non-standardized and
potentially confusing to users of the search form.

We remove the `fetchingExtendedOptions` flag from the `disable` prop since we no
longer need to wait for options to be loaded.

The value for `placeholder` was modeled after the text search field.