Use CreatableInputOnly for product types of claim form
This change mirrors the change made to the facility search sidebar. Now that we
are no longer populating a drop down with product types we remove the choices
from claim detail serializer.

The default value of the new `selectPlaceholder` prop was copied from the value
that appeared before the change.