Exclude details from facility list API by default
Serializing contributor fields and extended fields requires multiple database
lookups and is expensive. We add a `detail` query parameter to the facilities
list and only include the `contributors`, `contributor_fields`, and
`extended_fields` in the serialized output if that `detail` parameter is set to
true.

This change creates a regressing in the CSV downloads because the additional
fields are no longer available. That should be fixed in a future commit.