Preserve value of parent company search field on refresh
Previously we were clearing out any values that aren't in the contributors
dropdown on refresh, but since we now allow any value in the parent company
search box we don't want that. This section also has the purpose of mapping IDs
to names, but since we're not using the dropdown any IDs will have been
explicitly entered by users and probably shouldn't be overwritten anyway.