Update search text styles and copy
The OAR team prefers the search text about extended fields to match
in the two locations on the sidebar, so it has been updated.

Font sizes for have been updated to more closely match
the wireframes. Note that the 'Need tips for searching the OAR?'
link is a lavender in the wireframes; however, for the sake of embedded
maps, we need to keep colored text consistent, so this has been made
the standard primary color found in the config.