Fix handling of numeric custom cells in Excel
Custom cells can have numeric values. When parsing the sheets, a value
that appears as 130 in the workbook can appear as a float 130.0.

This commit attempts to fix this issue by using an
int(float(x)) nested cast to ensure that the decimal portion is removed.

Note that this will convert all floats to integers when submitted as
custom text. If a contributor would like to submit floats in a custom
text field in the future, we will need to implement that as an
enhancement.