Add a helper sciprt to run tile load test infinitely
The tile load test uses virtual users and iterations to chunk and run requests
read from a HAR file and exists after one set of tile requests is done. This
helper loops a run of the scripts with a different CACHE_KEY_SUFFIX for each run
to allow for testing sustained continuous load.