Add a homepage browsing load test script
This script was initially created by browsing the home page, exporting the
requests as a HAR file, converting the HAR file to a script using k6.io, then
modifying the script to simulate the parallel request behavior of a browser.

Making API requests requires a CLIENT_KEY so we update the load testing compose
file to pass through an appropriate environment variable.