Revise tests to accommodate exact match
We use exact match instead of fuzzy matching for parent_company for now,
so tests are revised to make sure the ParentCompanyTestCase input is
exactly the same as the contributor (aka. test contributor 1).
In test_filter_by_name_verified, we suspect that not using `similarity`
changes the sorting of same names, making the second name comes at
first when both names are the same. So we mark c1 as verified to test if
sorting by verified is working as expected.