Add extended field reports
- Contributors With Extended Fields Cumulative: This report lists
the number of contributors who have submitted each field type. It is
organized into months, where each month's count includes all extended
fields with an update date of that month or earlier.
- Facilities With Extended Fields Cumulative: This report lists
the number of facilities that have each field type. It is
organized into months, where each month's count includes all extended
fields with an update date of that month or earlier.
- Processing Type Facility Type Matched Values: This report lists each
taxonomy value that has been assigned to an extended field, accompanied
by a list of the raw values that have been matched to that value.
- Processing Type Facility Type Unmatched Values: This report lists
each raw value that was submitted as a Processing_Type_Facility_Type
and wasn't matched to a taxonomy value, as well as a count of the number
of times that raw value was submitted.
- Submitted Product Type Values: This report lists each raw value that
was submitted for a product_type extended field, as well as a count of
the number of times that raw value was submitted.