Add confirm/reject report
The OAR requested the ability to compare the number of open vs completed potential
match items. If all the potential matches for an item are rejected, a new
facility is created and the status of the item is set to `CONFIRMED_MATCH`, so
we don't need special code to cover rejection vs confirmation.