Include project name in Terraform resource names
In places where resource names were differentiated only by `var.environment` we
need to also include the project name since both OAR and OSH staging will exist
in the same account.

The `aws_batch.py` module was using string interpolation to build the name of an
AWS Batch job queue and job definition to be used for submitting jobs based on
the environment. To keep the construction of resource names exclusively within
Terraform we now pass the names through to the Django app via environment
variables