Integrate claims with main fields
It is important to indicate to users which items come from claims,
but now that several claim fields are included elsewhere in the
facility details, having a separate claims section caused duplication
of data presented. To handle this, the claims fields are no longer
separated into a separate area, but instead items from claims are
indicated with a badge, and tooltips contain the explanatory text.