Fix VectorGrid bug
We use two layers to display facilities in the map - the
VectorTileFacilitiesLayer and the VectorTileFacilityGridLayer.
The VectorTileFacilityGridLayer is the 'dots' on the map, and
is shown when the map is zoomed out to a certain level.

When the map was zoomed in far enough that the grid is hidden,
and the user navigated away from the map, an error was being thrown.
By only rendering the grid layer when it's visible, that error is
prevented.