Update sector indexing and add a sector indexing command
In a previews commit we updated facility detail serialization to include sectors
from inactive or private items. This commit extends that change to the facility
indexing process, introducing similar filtering and sorting changes to ensure
that the index is updated with the correct sectors.

Since we are updating the way the list of sectors is calculated for a facility
we need to update all sector values in the index. We add a new `index_sectors`
management command do a targeted update of the sector column.

We update a unit test because we are now considering multiple submissions from
the same contributor to be updates and only use the most recent submission to
get the list of sectors to index. Without this change there would be no way for
a contributor to correct a misaken sector submission.