Refactor index_sectors can call it after a faiclity is POSTed
When a facility is submitted via API the facility list item status is changed to
from `GEOCODED` to `MATCHED` after the facility has been indexed via post_save
signals on other models. Because the sector indexing is dependant on the list
item status we need to make sure to explicitly update the sector index values if
the item has been associated with a facility.