Update facility download format and move to Django
Updates the facility download format to split facility list items to
separate rows, with each facility having one base row and 0 or more
facility list item rows. In cases where a claim exists, the base row
is listed as contributed by the claimant; in other cases, the facility
created_from contributor is cited.

To simplify the process and decrease the amount of data fetched, the
formatting of the data into a CSV/XLSX-friendly format has been moved
to the backend.