Add Event model
Adds the Event model and corresponding migration. This new model will
be the foundation of the change notification system and a revised
history API. It makes use of contenttypes to allow us to use a generic
foreign field and to delete the associated objects without losing
the Event object or the ID of the deleted object.