Update rebuild_gazetteer to reflect model training changes
If the train_model flag is passed to the rebuild_gazetteer method, or if the database contains no
models, a new model is trained. Checking to see if the database contains no models should primarily
be used for the resetdb script.

If neither of these conditions are reached, it calls the load_gazetteer method to load the saved model
with the most recent timestamp.