Update sectors in FacilityIndex after upload
We previously didn't update the entire FacilityIndex
after a FacilityListItem was saved, only the custom
text. In most cases, the post_save for the FacilityMatch
covered the needed updates to the index. However,
in certain cases, the FacilityListItem for a new
facility is saved after the FacilityMatch is saved.
In these cases, when the Match.post_save updates the
FacilityIndex, the FacilityListItem's status is not yet
MATCHED or CONFIRMED_MATCH, so fields such as sector that
are only added to the index if a FacilityListItem with one of
those statuses exists are excluded.

Adding indexing to the FacilityListItem.post_save when the
item is in a valid status results in the sectors being
added to the index as expected.