Update Adjust Facility Matches dashboard
Admin users have had a number of instances where they attempted to
adjust a match but were unable to due to the current status of the
match. This shows additional information about the matches to assist
admins in using the tool, and also disables the adjustment buttons
for certain match statuses.