Display warning for missing claim locations
Claim addresses from before the claim.facility_location was added
might in some cases be impossible to geocode. In this case, the claim
address will still be prioritized, but the location coordinates will
be set to the original facility address. In this case, a warning has
been added to the GPS to inform users that the address might not
match the displayed location.

Note that this situation should only impact a small number of existing
claims, as many claim addresses are likely to be geocodable, and new
claims will not be accepted with invalid addresses.