Add AWS Batch resources for delivering notifications
Facility event notifications will be delivers by a management command run via
AWS Batch. The commit adds Terraform to create a new compute environment, job
definition and and job queue for running these notification jobs.

The configuration of these AWS Batch resources was copied from the exiting
"default" resources and modified.

We originally tries to set general purpose `t3` as the instance tyupe but
recieved an error message during provisioning that that type was not available,
so we copied the types from the default configuration. We set the resources to a
modest 1 vCPU and 512 MB of RAM because the job will only be running a few
database queries requiring a modest amount of data and making a single POST
request.