Fix facility details sector display
The sector date in the sidebar was attempting to render the date
based on the sector's created_at date, which is not sent from the
backend, resulting in the current date being displayed instead of the
correct date. We now use the updated_at date, which displays as
expected.

Sectors from claims be marked with the verified badge and prioritized
over other items, as we do for other fields.